app: Fix up editor router and move into editor module
